pyproject.toml
[project]
name = "course-llm-langchain"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "aaweaver-actuary", email = "andrewayersweaver+github@gmail.com" }
]
dependencies = [
    "langchain>=0.1.20",
    "transformers>=4.40.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.4.4",
    "ruff-lsp>=0.0.53",
    "langchain-cli>=0.0.22",
    "openai>=1.28.1",
    "langgraph>=0.0.48",
    "langserve[all]>=0.1.1",
    "langchain-openai>=0.1.6",
    "llama-cpp-python>=0.2.73",
    "huggingface-hub>=0.23.0",
    "huggingface-cli>=0.1",
    "ipykernel>=6.29.4",
    "jupyter>=1.0.0",
    "notebook>=7.1.3",
    "langchain-chroma>=0.1.0",
    "beautifulsoup4>=4.12.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/course_llm_langchain"]
